+++ src/views/Pages/PrivatePages/Dashboard/UserInfo/UserInfo.js
@@ -1,9 +1,10 @@
-import React from 'react';
+import React, { useEffect, useFetch } from 'react';
 import { useState } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 import { updateUserBio } from '~/redux/boarding/action';
 import { Button, Header, Checkbox, Modal } from 'semantic-ui-react';
 import { updateAvailability } from '~/redux/boarding/action';
+import { checkAvailability } from '~/redux/boarding/action';
 
 
 function UserInfo() {
@@ -12,35 +13,45 @@ function UserInfo() {
     const [openBioPopUp, setOpenBioPopUp] = React.useState(false);
     const [openAvailability, setOpenAvailability] = React.useState(false);
     const [switchAvailability, setSwitchAvailability] = React.useState(true);
-    const [availabilityDuration, setAvailabilityDuration] = React.useState({value:'30 minutes'});
+    const [availabilityDuration, setAvailabilityDuration] = React.useState({ value: '30 minutes' });
     const userInfo = useSelector(state => state.auth.user);
-    // const { user: { meeter_meet_slug } } = user;
-    console.log(userInfo);
+     
+ 
+    useEffect(() => { 
+        dispatch(checkAvailability(userInfo.id));  
+    }, [])
+ 
+
     const [editBio, setEditBio] = useState({ value: userInfo.meeter_bio });
-    // const [userInfo, setUserInfo] = useState(JSON.parse(localStorage.getItem('userInfo')));
     const openAvailabilityPopup = (value) => {
         if (value == false) {
-            setOpenAvailability(true); 
+            setOpenAvailability(true);
         } else {
-            setOpenAvailability(false); 
+            setOpenAvailability(false);
+            const data = {
+                "meeter_availibility": "no"
+            }
+            console.log(data);
+            dispatch(updateAvailability(data));
         }
     }
     const changeAvailability = () => {
         console.log(availabilityDuration);
-        if (openAvailability == true) { 
+        if (openAvailability == true) {
             const data = {
                 "meeter_availibility": "yes",
                 "time_duration": availabilityDuration.value
             };
             console.log(data);
-             dispatch(updateAvailability(data));
-        } else { 
+            dispatch(updateAvailability(data));
+        } else {
             const data = {
                 "meeter_availibility": "no"
             }
             console.log(data);
-             dispatch(updateAvailability(data));
+            dispatch(updateAvailability(data));
         }
+        setOpenAvailability(false);
     }
     const updateBio = () => {
         const data = {
@@ -56,12 +67,12 @@ function UserInfo() {
         console.log(e);
         if (e == false) {
             setSwitchAvailability(true);
-        }else{
+        } else {
             setSwitchAvailability(false);
         }
     };
-    const changeAvailabilityDuration  = (value) => {
-        setAvailabilityDuration({value:value});
+    const changeAvailabilityDuration = (value) => {
+        setAvailabilityDuration({ value: value });
     };
     return (
         <>
@@ -191,7 +202,7 @@ function UserInfo() {
                                                     <span className="small-size">Your Availability Status</span>
                                                     <div className="align-self-center row d-flex mx-0 my-2">
                                                         <div className="button b2" id="button-10">
-                                                            <input id="modalAvilabilityChk" onChange={e=>switchAvailabitly(e.target.checked)} type="checkbox" className="checkbox w-100 h-100 p-0 m-0" />
+                                                            <input id="modalAvilabilityChk" onChange={e => switchAvailabitly(e.target.checked)} type="checkbox" className="checkbox w-100 h-100 p-0 m-0" />
                                                             <div className="knobs small-size">
                                                                 <span className="small-size">Available</span>
                                                             </div>
@@ -199,86 +210,86 @@ function UserInfo() {
                                                     </div>
                                                 </div>
                                                 {switchAvailability == true && <div>
-                                                <div className="small-size text-left opacity-8 py-2">
-                                                    When your status is set to ‘Available’, your Requestors and You will receive browser and email notifications in order to start your meetings.
+                                                    <div className="small-size text-left opacity-8 py-2">
+                                                        When your status is set to ‘Available’, your Requestors and You will receive browser and email notifications in order to start your meetings.
                                                 </div>
-                                                <div className="form-group mb-0 px-0 py-3">
-                                                    <h5 className="small-size align-self-center text-left mb-2 pb-1">I am Available for the next:</h5>
-                                                    <div className="radio">
-                                                        <input 
-                                                        id="radio-1" 
-                                                        name="radio" 
-                                                        type="radio"
-                                                        value="10 minutes"
-                                                        onChange={e => changeAvailabilityDuration(e.target.value)}
-                                                         />
-                                                        <label for="radio-1" className="radio-label small-size opacity-8">10 minutes</label>
+                                                    <div className="form-group mb-0 px-0 py-3">
+                                                        <h5 className="small-size align-self-center text-left mb-2 pb-1">I am Available for the next:</h5>
+                                                        <div className="radio">
+                                                            <input
+                                                                id="radio-1"
+                                                                name="radio"
+                                                                type="radio"
+                                                                value="10 minutes"
+                                                                onChange={e => changeAvailabilityDuration(e.target.value)}
+                                                            />
+                                                            <label for="radio-1" className="radio-label small-size opacity-8">10 minutes</label>
+                                                        </div>
+                                                        <div className="radio">
+                                                            <input
+                                                                id="radio-2"
+                                                                name="radio"
+                                                                type="radio"
+                                                                value="30 minutes"
+                                                                onChange={e => changeAvailabilityDuration(e.target.value)}
+                                                            />
+                                                            <label for="radio-2" className="radio-label small-size opacity-8">30 minutes</label>
+                                                        </div>
+                                                        <div className="radio">
+                                                            <input id="radio-3"
+                                                                name="radio"
+                                                                type="radio"
+                                                                value="1 hour"
+                                                                onChange={e => changeAvailabilityDuration(e.target.value)}
+                                                            />
+                                                            <label for="radio-3" className="radio-label small-size opacity-8">1 hour</label>
+                                                        </div>
+                                                        <div className="radio">
+                                                            <input id="radio-4"
+                                                                name="radio"
+                                                                type="radio"
+                                                                value="2 hours"
+                                                                onChange={e => changeAvailabilityDuration(e.target.value)}
+                                                            />
+                                                            <label for="radio-4" className="radio-label small-size opacity-8">2 hours</label>
+                                                        </div>
+                                                        <div className="radio">
+                                                            <input id="radio-5"
+                                                                name="radio"
+                                                                type="radio"
+                                                                value="4 hours"
+                                                                onChange={e => changeAvailabilityDuration(e.target.value)}
+                                                            />
+                                                            <label for="radio-5" className="radio-label small-size opacity-8">4 hours</label>
+                                                        </div>
+                                                        <div className="radio">
+                                                            <input id="radio-6"
+                                                                name="radio"
+                                                                type="radio"
+                                                                value="8 hours"
+                                                                onChange={e => changeAvailabilityDuration(e.target.value)}
+                                                            />
+                                                            <label for="radio-6" className="radio-label small-size opacity-8">8 hours</label>
+                                                        </div>
                                                     </div>
-                                                    <div className="radio">
-                                                        <input 
-                                                        id="radio-2"
-                                                        name="radio"
-                                                        type="radio"
-                                                        value="30 minutes" 
-                                                        onChange={e => changeAvailabilityDuration(e.target.value)}
-                                                         />
-                                                        <label for="radio-2" className="radio-label small-size opacity-8">30 minutes</label>
+                                                    <div className="MeeterLine"></div>
+                                                    <div className="form-group px-0 py-3 mb-0">
+                                                        <div className="chkbox">
+                                                            <label>
+                                                                <input type="checkbox" name="checkbox" value="css" />
+                                                                <span className="small-size opacity-8">Open Meeting Room Window</span>
+                                                            </label>
+                                                        </div>
+                                                        <div className="chkbox">
+                                                            <label>
+                                                                <input type="checkbox" name="checkbox" value="css" checked />
+                                                                <span className="small-size opacity-8">Send Notifications to all Requestors</span>
+                                                            </label>
+                                                        </div>
                                                     </div>
-                                                    <div className="radio">
-                                                        <input id="radio-3" 
-                                                        name="radio" 
-                                                        type="radio" 
-                                                        value="1 hour" 
-                                                        onChange={e => changeAvailabilityDuration(e.target.value)}
-                                                        />
-                                                        <label for="radio-3" className="radio-label small-size opacity-8">1 hour</label>
-                                                    </div>
-                                                    <div className="radio">
-                                                        <input id="radio-4" 
-                                                        name="radio" 
-                                                        type="radio" 
-                                                        value="2 hours" 
-                                                        onChange={e => changeAvailabilityDuration(e.target.value)}
-                                                        />
-                                                        <label for="radio-4" className="radio-label small-size opacity-8">2 hours</label>
-                                                    </div>
-                                                    <div className="radio">
-                                                        <input id="radio-5" 
-                                                        name="radio" 
-                                                        type="radio" 
-                                                        value="4 hours" 
-                                                        onChange={e => changeAvailabilityDuration(e.target.value)}
-                                                        />
-                                                        <label for="radio-5" className="radio-label small-size opacity-8">4 hours</label>
-                                                    </div>
-                                                    <div className="radio">
-                                                        <input id="radio-6" 
-                                                        name="radio" 
-                                                        type="radio" 
-                                                        value="8 hours" 
-                                                        onChange={e => changeAvailabilityDuration(e.target.value)}
-                                                        />
-                                                        <label for="radio-6" className="radio-label small-size opacity-8">8 hours</label>
-                                                    </div>
-                                                </div>
-                                                <div className="MeeterLine"></div>
-                                                <div className="form-group px-0 py-3 mb-0">
-                                                    <div className="chkbox">
-                                                        <label>
-                                                            <input type="checkbox" name="checkbox" value="css" />
-                                                            <span className="small-size opacity-8">Open Meeting Room Window</span>
-                                                        </label>
-                                                    </div>
-                                                    <div className="chkbox">
-                                                        <label>
-                                                            <input type="checkbox" name="checkbox" value="css" checked />
-                                                            <span className="small-size opacity-8">Send Notifications to all Requestors</span>
-                                                        </label>
-                                                    </div>
-                                                </div>
-                                                <div class="MeeterLine"></div>
+                                                    <div class="MeeterLine"></div>
                                                 </div>}
-                                            </div> 
+                                            </div>
                                         </Modal.Description>
                                     </Modal.Content>
                                     <Modal.Actions>
